20160129_034150 Update - [Rake] Enable Rake tasks under 'server:sqale' only on Sqale

+++ lib/tasks/server/sqale.rake
@@ -15,24 +15,39 @@ namespace :server do
       # @see https://gist.github.com/banyan/4943864
       desc "Process 'upload' directory on 'postinstall'"
       task :process_upload_dir do
+        begin
+          raise Sqale::EnvironmentError unless Rails.application.on_sqale?
 
-        path = "#{ENV['HOME']}/current/public/uploads"
+          path = "#{ENV['HOME']}/current/public/uploads"
 
-        if File.directory?(path)
-          FileUtils.mv(path, "#{path}.old")
-        elsif File.symlink?(path)
-          FileUtils.rm(path)
+          if File.directory?(path)
+            FileUtils.mv(path, "#{path}.old")
+          elsif File.symlink?(path)
+            FileUtils.rm(path)
+          end
+
+          FileUtils.symlink("#{ENV['HOME']}/uploads", path)
+
+        rescue Sqale::EnvironmentError => e
+          puts e.class
+          nil
         end
-
-        FileUtils.symlink("#{ENV['HOME']}/uploads", path)
       end
 
       # 'Whenever' による Crontab の設定
       desc 'Update Cron settings'
       task :update_cron do
+        begin
+          raise Sqale::EnvironmentError unless Rails.application.on_sqale?
 
-        Dir.chdir("#{ENV['HOME']}/current")
-        system('bundle exec whenever --update-crontab')
+          Dir.chdir("#{ENV['HOME']}/current")
+          system('bundle exec whenever --update-crontab')
+
+        rescue Sqale::EnvironmentError => e
+          puts e.class
+          nil
+        end
+
       end
 
     end
@@ -47,8 +62,8 @@ namespace :server do
             File.delete(file)
           end
 
-        rescue Sqale::EnvironmentError
-          puts 'Sqale::EnvironmentError'
+        rescue Sqale::EnvironmentError => e
+          puts e.class
           nil
         end
       end
